pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "flask>=3.1.2",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
